backend/requirements.txt
# filepath: c:\Users\cadsa\OneDrive\Personal\Full stack web development\wellnessCentre\backend\requirements.txt
Django==5.0.14
djangorestframework==3.15.2
djangorestframework-simplejwt==5.4.0
dj-database-url==2.1.0
psycopg2-binary==2.9.9
python-decouple==3.8
Pillow==10.2.0
drf-yasg==1.21.7
django-cors-headers==4.3.1
celery==5.3.6
redis==5.0.1
django-celery-beat==2.5.0
django-celery-results==2.5.1
django-storages==1.14.2
coverage==7.5.1
pytest==8.0.0
pytest-django==4.8.0
pytest-cov==4.1.0
WeasyPrint==60.1
whitenoise==6.6.0
gunicorn==21.2.0
stripe==8.1.0
Faker==24.5.0
django-filter==24.1
twilio==9.0.0
requests==2.31.0
pyjwt==2.8.0
django-health-check==3.17.0
pandas==2.2.2
matplotlib==3.8.3
